move and rename hook file
